BillBook
Developed an Android application for the Customers to add, update, view
and delete the bills with image upload feature. Designed interactive
prototypes and UI specifications, including screen layouts and user
interface elements. Provided the features such as finding the way to the
store(using Google Maps), contact the store(with phone call feature).